+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Extend cookie scrubbing to response cookies. ([#6391](https://github.com/getsentry/relay/pull/6391))
 
+**Internal**:
+
+- Add a flush timeout to the envelope buffer's in-memory queues. ([#6375](https://github.com/getsentry/relay/pull/6375))
+
 ## 26.9.0
 
 **Breaking Changes**: